Add Gate User Answer For Update And Delete Answer

+++ app/Http/Controllers/Api/V1/Thread/AnswerController.php
@@ -7,6 +7,7 @@ use App\Http\Repositories\AnswerRepository;
 use App\Models\Answer;
 use App\Models\Thread;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Symfony\Component\HttpFoundation\Response;
 
 class AnswerController extends Controller
@@ -36,19 +37,28 @@ class AnswerController extends Controller
         $request->validate([
             'content' => 'required',
         ]);
+        if (Gate::forUser(auth()->user())->allows('user_answer', $answer)) {
+            resolve(AnswerRepository::class)->update($request, $answer);
 
-        resolve(AnswerRepository::class)->update($request, $answer);
-
+            return \response()->json([
+                'message' => 'answer updated successfully'
+            ], Response::HTTP_OK);
+        }
         return \response()->json([
-            'message' => 'answer updated successfully'
-        ], Response::HTTP_OK);
+            'message' => 'Access denied'
+        ], Response::HTTP_FORBIDDEN);
     }
 
     public function destroy(Answer $answer)
     {
+        if (Gate::forUser(auth()->user())->allows('user_answer',$answer)){
         resolve(AnswerRepository::class)->destroy($answer);
         return \response()->json([
             'message' => 'answer deleted successfully'
         ], Response::HTTP_OK);
+        }
+        return \response()->json([
+            'message' => 'Access denied'
+        ],Response::HTTP_FORBIDDEN);
     }
 }
